requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
sse-starlette==2.1.0
python-dotenv==1.0.1